requirements.txt
acme==1.29.0
asgiref==3.5.2
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
bcrypt==3.2.2
beautifulsoup4==4.11.1
bootstrap4==0.1.0
boto3==1.24.36
botocore==1.27.36
Brotli==1.0.9
certbot==1.29.0
certifi==2022.6.15
cffi==1.15.1
channels==2.4.0
charset-normalizer==2.1.0
click==8.1.3
ConfigArgParse==1.5.3
configobj==5.0.6
constantly==15.1.0
cryptography==37.0.4
daphne==2.5.0
dash==2.6.0
dash-bootstrap-components==0.13.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
distro==1.7.0
dj-database-url==0.5.0
Django==3.2.14
django-bootstrap==0.2.4
django-bootstrap4==22.1
django-on-heroku==1.1.2
django-plotly-dash==2.0.0
django-storages==1.12.3
dpd-components==0.1.0
Flask==2.1.3
Flask-Compress==1.12
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
josepy==1.13.0
letsencrypt==0.7.0
MarkupSafe==2.1.1
numpy==1.23.1
paramiko==2.11.0
parsedatetime==2.6
plotly==5.9.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyRFC3339==1.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.1
requests==2.28.1
requests-toolbelt==0.9.1
s3transfer==0.6.0
scipy==1.8.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
storage==0.0.4.3
tenacity==8.0.1
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
unicorn==2.0.0
urllib3==1.26.10
Werkzeug==2.0.3
whitenoise==6.2.0
zope.component==5.0.1
zope.event==4.5.0
zope.hookable==5.1.0
zope.interface==5.4.0